Fix buffer over-read in json_decode
json_i_decodeElement uses sscanf() to parse floats. Unfortunately
sscanf() expects to be passed a NULL-terminated string, and so calls
strlen()internally. This causes a buffer-overread in the case that the
buffer is not NULL-terminated (e.g. when it's from [NSData
dataWithContentsOfFile:])

The fix is simple, we just copy floats into a temporary buffer before
parsing them.