Initialize KSCrashMonitor_CPPException stackCursor
Our __cxa_throw override is only called if the exception is thrown
from the same image as our __cxa_throw lives in, e.g. the main app
if KSCrash is linked in as a static library.

In cases where it's not called the advanceCursor and resetCursor
members of the KSStackCursor are 0, resulting in a crash during
crash handling.

By always initializing the stackCursor we at least ensure that we
don't crash during reporting, even if the result is an empty
stack trace for that thread.