Deduplicate logging to prepare for logging override API
The 'basic' versions of the C and ObjC loggers have been retired,
in place of just having the macro fill in default arguments for
the full function.

The Objective-C full function now only takes care of converting the
CFStringRef into a C string, that is then passed on to the lower
level function i_kslog_log().

The same goes for i_kslog_logC, which prepares the va-args, and
forwards to i_kslog_log(). This is then the only place where the
the level, file, line, and function are formatted and log output
is written, which should make the code easier to maintain, and
allows inserting API entrypoints in i_kslog_log().