Add C API to allow clients to react to the report being finished writing
The existing kscrash_setCrashNotifyCallback() is called during report
writing, and is meant to be used for adding user info to the report.

To be able to e.g. spawn a new process that reads and sends the finished
report, we need a callback that happens after the report has been fully
written.