fix: prevent possible overflow
This was caught by an automated security check in GitHub: multiplying
two 32-bit values could result in overflow. On 32 bit systems, the
result may be truncated anyways, but on 64 bit, the void* pointer will
still be able to hold the full value as long as it can fit in a
`uintptr_t`, which is always an unsigned long.