Invert dependency from individual monitors to KSCrashMonitor.c, using DI instead.
All KSCrashMonitorAPI structs must now be initialized using kscm_initAPI() before they are filled out. This ensures that there are always default implementations and none of the function pointers are NULL. Since no pointers can be accidentally null, functions such as "kscm_getMonitorId()" are no longer necessary, and have been removed.

KSCrashMonitorAPI now has a new function "init", which is used to pass in "notify" and "handle" callbacks. These replace the functions "kscm_notifyFatalExceptionCaptured", "kscm_notifyNonFatalExceptionCaptured", and "kscm_handleException".

Information about things such as async safety and "fatalness" of a crash are now encoded in KSCrash_ExceptionHandlingPolicy, which is used both as a recommendation from a monitor instance to KSCrashMonitor, and also as a way to advertise what KSCrashMonitor has decided will be the final policy. This will be expanded in a future commit.